+++ src/channels/index.mjs
@@ -16,5 +16,6 @@ import * as hijriah from "./hijriah.mjs";
 import * as pesawat from "./pesawat.mjs";
 import * as doa from "./doa.mjs";
 import * as wiki from "./wiki.mjs";
+import * as maritim from "./maritim.mjs";
 
-export const CHANNELS = { gempa, cuaca, berita, wilayah, kodepos, kurs, saham, ihsg, sholat, quran, hijriah, pesawat, doa, wiki, data, bps, sosial, udara };
+export const CHANNELS = { gempa, cuaca, berita, wilayah, kodepos, kurs, saham, ihsg, sholat, quran, hijriah, pesawat, doa, wiki, maritim, data, bps, sosial, udara };