+++ test.sh
@@ -32,6 +32,7 @@ check "hijriah"  '"hijriah"'     $J hijriah
 check "pesawat"  '"total_di_wilayah"' $J pesawat --limit 3
 check "doa"      '"arab"'        $J doa
 check "wiki"     '"ringkasan"'   $J wiki soekarno
+check "maritim"  '"tinggi_gelombang"' $J maritim "selat sunda"
 check "gempa-live" '"magnitudo"' $J gempa --live
 check "cuaca"    '"lokasi"'     $J cuaca jasinga bogor
 check "bps"      'webapi.bps.go.id\|"indikator"' $J bps inflasi